--- src/tutorials/gno.land/gbe/chain/03-events/gno.mod
@@ -1,1 +1,0 @@
-module gno.land/r/demo/event
