+++ src/tutorials/gno.land/gbe/chain/04-call_stack/callstack_test.gno
@@ -1,0 +1,9 @@
+package event
+
+import (
+	"testing"
+)
+
+func TestExample(t *testing.T) {
+	println(Example())
+}
