--- src/tutorials/gno.land/gbe/chain/05-banker/gno.mod
@@ -1,1 +1,0 @@
-module gno.land/r/demo/event
