+++ src/tutorials/gno.land/gbe/chain/05-coins/coins1.gno
