+++ src/tutorials/gno.land/gbe/chain/05-coins/coins1_test.gno
